Use alternative API for virtual paths
This effectively replaces calls to fugitive#chmod() with calls to
g:io_fugitive.chmod().  I added this alternative API a while back to
eliminate the need to load the autoload file.  Recent changes to nvim
break our use of 'runtimepath' here, so this is as good of time to
switch over as any.