Re-detect 'filetype' after typing a shebang line
Also, normalize `#!cmd` to `#!/usr/bin/env cmd`.  And for a blank
shebang of `#!`, do the opposite and provide a default based on
'filetype'.

Bonus undocumented feature:  Delete the obnoxious comment character that
gets inserted on the subsequent line when 'fo' contains "r" and
'comments' contains ":#" (several file types annoyingly default to this
combo).