Account for guioptions=! in :SudoEdit/:SudoWrite
This was the last straw for `tee >/dev/null`.  Turns out it's very easy
to replace it with a small shell script and avoid the stdout redirect.
This works with both old versions of sudo which split SUDO_EDITOR on
spaces as well as new ones that handle quotes (our own quotes are
ignored).