Fix redundant argument expansion, second attempt
In 20f11a6693ef6bc14aaac17da368710d60615a70, I attempted to fix this by
changing -nargs=1 to -nargs=+, based on the observation that it seemed
to prevent Vim from implicitly expanding <q-args> before handing it to
us.  Upon further investigation, that actually gives us a partially
expanded mess, expanding "%" but not "$VAR".  If you can't beat 'em,
join 'em.  This commit reverts to -nargs=1, and drops our own calls to
expand() entirely.

This also restores the zero argument version of :SudoEdit, which was
accidentally removed in the previous attempt.

Resolves: https://github.com/tpope/vim-eunuch/issues/101